requirements.txt
streamlit==1.28.0          # ← keep newest
opencv-python==4.8.1.78    # GUI build; remove headless
ultralytics==8.0.196       # keep newest
numpy==1.26.0              # matches pandas wheels
pandas==2.1.1
psycopg2-binary==2.9.7
python-dotenv==1.0.0
PyYAML==6.0.1
face-recognition==1.3.0
boto3==1.28.65
langchain==0.0.346
openai==0.28.1
streamlit-chat==0.1.0
pillow==10.0.1
tensorboard==2.14.1
